pyproject.toml
[tool.poetry]
name = "roompricegenie"
version = "0.1.0"
description = ""
authors = ["Sirojiddin Yakubov <yakubov9791999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
amqp = "5.2.0"
asgiref = "3.7.2"
billiard = "4.2.0"
celery = "5.3.6"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
click = "8.1.7"
click-didyoumean = "0.3.0"
click-plugins = "1.1.1"
click-repl = "0.3.0"
colorama = "0.4.6"
cron-descriptor = "1.4.3"
django = "5.0.3"
django-celery-beat = "2.6.0"
django-filter = "24.1"
django-timezone-field = "6.1.0"
djangorestframework = "3.15.0"
drf-yasg = "1.21.7"
faker = "24.3.0"
idna = "3.6"
inflection = "0.5.1"
iniconfig = "2.0.0"
kombu = "5.3.5"
packaging = "24.0"
pathlib = "1.0.1"
pluggy = "1.4.0"
prompt-toolkit = "3.0.43"
psycopg2-binary = "2.9.9"
python-crontab = "3.0.0"
python-dateutil = "2.9.0.post0"
pytz = "2024.1"
pyyaml = "6.0.1"
redis = "5.0.3"
requests = "2.31.0"
six = "1.16.0"
sqlparse = "0.4.4"
tzdata = "2024.1"
uritemplate = "4.1.1"
urllib3 = "2.2.1"
vine = "5.1.0"
wcwidth = "0.2.13"
gunicorn = "^21.2.0"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
